Modify Setting

+++ app/Http/Controllers/Admin/SettingController.php
@@ -5,7 +5,9 @@ namespace App\Http\Controllers\Admin;
 use App\Http\Controllers\Controller;
 use App\Http\Requests\SettingUpdateRequest;
 use App\Models\Setting;
+use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Storage;
+use Illuminate\Validation\Rule;
 
 class SettingController extends Controller
 {
@@ -14,11 +16,34 @@ class SettingController extends Controller
 		return view("Admin.setting.index");
 	}
 
+  public function store(Request $request)
+  {
+    $request->validate([
+      'group' => ['required', Rule::in(['general', 'social'])],
+      'label' => ['required', 'string', 'max:191'],
+      'name' => ['required', 'string', 'max:191', 'alpha_dash:ascii'],
+      'type' => ['required', 'string', Rule::in(['text', 'textarea', 'image'])],
+    ]);
+
+    Setting::create([
+      'group' => $request->group,
+      'label' => $request->label,
+      'name' => $request->name,
+      'type' => $request->type,
+      'value' => '',
+    ]);
+
+    toastr()->success('تنظیمات با موفقیت بروزرسانی شد.');
+    return redirect()->back();
+  }
+
 	public function edit(String $group)
 	{
 		$settingTypes = Setting::query()->where("group", $group)->get()->groupBy("type");
-		$group = Setting::GROUP[$group];
-		return view("Admin.setting.edit", compact(["settingTypes", "group"]));
+    $groupName = $group;
+    $group = Setting::GROUP[$group];
+
+		return view("Admin.setting.edit", compact(["settingTypes", "group", 'groupName']));
 	}
 
 	public function update(SettingUpdateRequest $request)
@@ -26,7 +51,7 @@ class SettingController extends Controller
 		$inputs = $request->except(['_token', '_method']);
 		foreach ($inputs as $name => $value) {
 			if ($setting = Setting::where('name', $name)->first()) {
-				if ($setting->type == 'image' || $request->file($name)->isValid()) {
+				if ($setting->type == 'image' && $request->file($name)->isValid()) {
 					if ($setting->value) {
 						Storage::delete($setting->value);
 					}
@@ -35,16 +60,19 @@ class SettingController extends Controller
 				$setting->update(['value' => $value]);
 			}
 		}
-		toastr()->success('تنظیات با موفقیت بروزرسانی شد.');
+		toastr()->success('تنظیمات با موفقیت بروزرسانی شد.');
 		return redirect()->back();
 	}
 
-	public function destroyFile(Setting $setting) 
+	public function destroyFile(Setting $setting)
 	{
 		if($setting->type !== 'image') {
 			toastr()->warning('تایپ فایل انتخاب شده برابر با عکس نیست.');
 		}else{
-			Storage::delete($setting->value);
+			Storage::disk('public')->delete($setting->value);
+      $setting->value = '';
+      $setting->save();
+
 			toastr()->success('فایل با موفقیت حذف شد.');
 		}
 		return redirect()->back();
